altera service

+++ back/src/main/java/com/pi/api/service/CategoryService.java
@@ -2,8 +2,12 @@ package com.pi.api.service;
 
 import com.pi.api.entity.Category;
 import com.pi.api.repository.CategoryRepository;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,28 +15,44 @@ import java.util.Optional;
 @Service
 public class CategoryService {
 
-    @Autowired
-    private CategoryRepository categoryRepository;
+	@Autowired
+	private CategoryRepository categoryRepository;
 
-    public Category salvar(Category category) {
-        return categoryRepository.save(category);
-    }
+	public Category salvar(Category category) {
+		return categoryRepository.save(category);
+	}
 
-    public Category atualizar(Category category){
-        return categoryRepository.save(category);
-    }
+	public boolean idExiste(Long id) {
+		return categoryRepository.existsById(id);
+	}
 
-    public void excluir(Long id) {
-        categoryRepository.deleteById(id);
-    }
+	public ResponseEntity<Category> atualizar(Category category){
 
-    public Optional<Category> buscarPorId(Long id) {
-        return categoryRepository.findById(id);
-    }
+		ResponseEntity<Category> response;
 
-    public List<Category> buscarTodos() {
-        return categoryRepository.findAll();
-    }
+		if (idExiste(category.getId())) {
+			response = ResponseEntity.status(HttpStatus.OK).body(categoryRepository.save(category));
+		} else {
+			response = ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+		}
 
+		return response; //NAO USAR RESPONSE ENTITY NO SERVICE
+	}
+
+	public Optional<Category> buscarPorId(Long id) {
+
+		return categoryRepository.findById(id);
+	}
+
+	public List<Category> buscarTodos() {
+		return categoryRepository.findAll();
+	}
+
+	public void excluir(Long id) {
+
+		if (idExiste(id)) {
+			categoryRepository.deleteById(id);
+		}
+	}
 
 }
